requirements.txt
asgiref==3.5.2
autopep8==1.7.0
black==22.3.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Django==4.0.5
djangorestframework==3.13.1
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.9.1
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
urllib3==1.26.11
